+++ cloutility/cloutapi.go
@@ -14,7 +14,7 @@
  * OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
  */
 
-package cloutapi
+package cloutility
 
 import (
 	"bytes"
