+++ examples/enroll/README.md
@@ -1,5 +1,5 @@
 # WorkstationEnroll example
 
-This is an example application based on the cloutapi-package to automatically enroll a backup-node as a linux Workstation with a predefined retention period and options. 
+This is an example application based on the cloutility-package to automatically enroll a backup-node as a linux Workstation with a predefined retention period and options.
 
 Example code is provided in order to demonstrate how a simple enrollment application could leverage the cloutility-api-client package to auto-enroll backup-nodes against Safespring's backup service.
